Added unit tests, coverage and cmake.

+++ MATLAB/example_1.m
@@ -195,7 +195,14 @@ gorilla_plot.write([path_RUN,'/gorilla_plot.inp']);
 tetra_grid.write([path_RUN,'/tetra_grid.inp']);
 
 %Create softlinks for used files
-! ln -s ../../../test_gorilla_main.x .
+if isfile('../../../test_gorilla_main.x')
+    ! ln -s ../../../test_gorilla_main.x .
+elseif isfile('../../../BUILD/SRC/test_gorilla_main.x')
+    ! ln -s ../../../BUILD/SRC/test_gorilla_main.x .
+else
+    disp('GORILLA not built, exiting the MatLab script')
+    return
+end
 ! ln -s ../../../MHD_EQUILIBRIA .
 
 %Run GORILLA code
